Initial commit for the warp-based operator interface
Operators now have a different interface based on whether they are meant
to be run on a GPGPU or on a CPU. In any case they have an open and a
close function. For the CPU they also have a consume(buffer_t) function.
For the GPGPU they must implement: consume_open(), consume_warp(T *,
unsigned int), consume_close().