Created a Makefile and implemented a HashJoin operator
Implemented a hashjoin (equijoin) operator and created a Makefile for
separate compilation. Also introduced an open/close pipeline and made a
small modification to the GPU operators' interface. A clean up was also
performed. The makefile uses a modified version of colornvcc
(https://github.com/kristi/colornvcc) for colorizing the output of nvcc.