BufferManager bug fixes
Buffer manager was not taking care of released (by the host) device
buffers which did not originate from the buffer manager. Also another
buf was fixed, where the device_buff was sized based on the
device_buff_size, while it has to also be at least keep_threshold.
Threadsafe_device_stack was also modified to be safe with regard to the
atomics/volatiles, but the changes were put back into comments, for when
they are required (theoretically, now it has some undefined behavior,
but it is a little bit faster; in practice, it currently works fine and
making the changes probably has no performance penalty, nevertheless,
they are commented out).